src/content_service/requirements.txt
psycopg2-binary==2.9.10
grpcio==1.71.0
grpcio-reflection==1.71.0
protobuf==5.29.4
confluent-kafka==2.9.0
